Renaming mainwindow to Login
