+++ BakeyPOS-Qt.pro
@@ -9,14 +9,14 @@ CONFIG += c++17
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-    main.cpp \
-    mainwindow.cpp
+    Login.cpp \
+    main.cpp
 
 HEADERS += \
-    mainwindow.h
+    Login.h
 
 FORMS += \
-    mainwindow.ui
+    Login.ui
 
 # Default rules for deployment.
 qnx: target.path = /tmp/$${TARGET}/bin
