+++ Login.h
@@ -1,0 +1,23 @@
+#ifndef LOGIN_H
+#define LOGIN_H
+
+#include <QMainWindow>
+
+QT_BEGIN_NAMESPACE
+namespace Ui {
+class MainWindow;
+}
+QT_END_NAMESPACE
+
+class Login : public QMainWindow
+{
+    Q_OBJECT
+
+public:
+    Login(QWidget *parent = nullptr);
+    ~Login();
+
+private:
+    Ui::MainWindow *ui;
+};
+#endif // LOGIN_H
