+++ main.cpp
@@ -1,10 +1,10 @@
-#include "mainwindow.h"
+#include "Login.h"
 #include <QApplication>
 
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-    MainWindow w;
+    Login w;
     w.show();  // ✅ This line is required to display the form
     return a.exec();
 }
